fix(backend): clean up query-builder — remove unused imports, move semantic helpers to semantic.ts
- Remove getOrderColumnExpression and ConditionWithConjunction from backend.ts imports (neither is used in the Backend class)
- Move computeSampleTimestamps and meanPoolEmbeddings from query-builder.ts to semantic.ts where they semantically belong
- Update backend.ts to import those two helpers from ./semantic instead of ./query-builder
- Remove the TODO comment referencing the move in query-builder.ts
