+++ src/backend/query-builder.ts
@@ -525,56 +525,3 @@ function applyExtraPropertyCondition(
   // Return the expression
   return eb('files.id', 'in', subquery);
 }
-
-// TODO: Move to src/backend/repositories/SemanticRepository.ts when that file is created.
-
-/**
- * Computes sample timestamps (in seconds) from a video duration and a list of ratios.
- */
-export function computeSampleTimestamps(
-  durationSeconds: number,
-  ratios: readonly number[],
-): number[] {
-  if (!Number.isFinite(durationSeconds) || durationSeconds <= 0) {
-    return [0, 1, 2, 3].slice(0, ratios.length);
-  }
-
-  const maxTimestamp = Math.max(0, durationSeconds - 0.05);
-  return ratios.map((ratio) => {
-    const normalizedRatio = Math.min(1, Math.max(0, ratio));
-    return Math.min(maxTimestamp, durationSeconds * normalizedRatio);
-  });
-}
-
-/**
- * Computes the mean-pooled (and L2-normalized) embedding from multiple frame embeddings.
- */
-export function meanPoolEmbeddings(embeddings: number[][]): number[] {
-  const first = embeddings[0];
-  const sum = first.slice();
-
-  for (let i = 1; i < embeddings.length; i++) {
-    const current = embeddings[i];
-    if (current.length !== sum.length) {
-      throw new Error('Cannot aggregate semantic video frame embeddings with mismatched sizes.');
-    }
-    for (let dim = 0; dim < sum.length; dim++) {
-      sum[dim] += current[dim];
-    }
-  }
-
-  for (let dim = 0; dim < sum.length; dim++) {
-    sum[dim] /= embeddings.length;
-  }
-
-  let norm = 0;
-  for (const value of sum) {
-    norm += value * value;
-  }
-  if (norm <= 0) {
-    return sum;
-  }
-
-  const scale = 1 / Math.sqrt(norm);
-  return sum.map((value) => value * scale);
-}
